+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Deprecate setting `babdev_pagerfanta.exceptions_strategy.out_of_range_page` and `babdev_pagerfanta.exceptions_strategy.not_valid_current_page` configuration options to any value, as of 3.0 they must be either "to_http_not_found" (default) or "custom"
 - Removed the dependency to TwigBundle and made Twig an optional dependency
 - Made the Symfony Translation component an optional dependency
+- Extracted the logic for building the route generator used by the Twig extension into `BabDev\PagerfantaBundle\RouteGenerator\RouteGeneratorFactoryInterface` and `BabDev\PagerfantaBundle\RouteGenerator\RouteGeneratorInterface` with default implementations for the existing generator
 
 ## 2.1.0 (2020-01-29)